requirements.txt
feedparser>=6.0
PyYAML>=6.0
openai>=1.0
pytz>=2024.1
requests>=2.30